WIP splitting out spi and i2c implementations
